Fix: database

+++ database/migrations/1768492104350_create_add_user_id_to_courses_table.ts
@@ -4,9 +4,11 @@ export default class extends BaseSchema {
   protected tableName = 'courses'
 
   async up() {
-    this.schema.alterTable(this.tableName, (table) => {
-      table.integer('user_id').unsigned().references('id').inTable('users').onDelete('SET NULL').nullable()
-    })
+    if (!(await this.schema.hasColumn(this.tableName, 'user_id'))) {
+      this.schema.alterTable(this.tableName, (table) => {
+        table.integer('user_id').unsigned().references('id').inTable('users').onDelete('SET NULL').nullable()
+      })
+    }
   }
 
   async down() {
