+++ database/migrations/1768559276304_create_guest_accesses_table.ts
@@ -4,13 +4,15 @@ export default class extends BaseSchema {
   protected tableName = 'guest_accesses'
 
   async up() {
-    this.schema.createTable(this.tableName, (table) => {
-      table.increments('id')
-      table.string('ip_address').notNullable()
-      table.integer('course_id').unsigned().references('courses.id').onDelete('CASCADE')
-      table.timestamp('created_at')
-      table.timestamp('updated_at')
-    })
+    if (!(await this.schema.hasTable(this.tableName))) {
+      this.schema.createTable(this.tableName, (table) => {
+        table.increments('id')
+        table.string('ip_address').notNullable()
+        table.integer('course_id').unsigned().references('courses.id').onDelete('CASCADE')
+        table.timestamp('created_at')
+        table.timestamp('updated_at')
+      })
+    }
   }
 
   async down() {
