+++ database/migrations/1768665136832_create_add_category_to_courses_table.ts
@@ -4,9 +4,11 @@ export default class extends BaseSchema {
   protected tableName = 'courses'
 
   async up() {
-    this.schema.alterTable(this.tableName, (table) => {
-      table.integer('category_id').unsigned().nullable().references('id').inTable('categories').onDelete('SET NULL')
-    })
+    if (!(await this.schema.hasColumn(this.tableName, 'category_id'))) {
+      this.schema.alterTable(this.tableName, (table) => {
+        table.integer('category_id').unsigned().nullable().references('id').inTable('categories').onDelete('SET NULL')
+      })
+    }
   }
 
   async down() {
